requirements.txt
folium==0.19.5
haversine==2.9.0
numpy==2.2.5
pandas==2.2.3
Pillow==11.2.1
plotly==6.0.1
streamlit==1.44.1
streamlit_folium==0.24.1
